[v0.90.1][runtime] Split runtime_v2.rs responsibilities (Closes #2207)
